move front-end related functions into  `deploy_frontend.go`
